This change allows json_encode functions to report
errors in a more helpful way.

	json_encode(['a' => ['b' => INF]])
	old message: strange double inf in function json_encode
	new message: /['a']['b']: strange double inf in function json_encode

This is helpful if json_encode is called with various data
from multiple places. With precise object path hint, users
can fix their code easier.

The performance impact is small (~1.35% for the worst case):

	name                   old time/op  new time/op  delta
	Json::EncodeSafeArray  1.52µs ± 1%  1.54µs ± 1%  +1.35%

This is due to the fact that vector/map loop is now split into
two parts, reducing the redundant branching. Without this optimization,
code becomes ~8% slower. It comes with a little bit of code duplication.

The object path stack is a fixed array of 8 `const char*`.
This array is allocated on a stack (part of the JsonEncoder object).
If object path is longer than 8 parts, it will be truncated like this:

	/['k1']['k2']['k3']['k4']['k5']['k6']['k7']['k8']...

All integer keys are stored as `.`:

	[9 => ['a' => INF]]  = /[.]['a']
	[10 => ['a' => INF]] = /[.]['a']

It's possible to increase the stack depth limit in the future.
8 -> 16 step is almost free.